requirements.txt
# Dev/Deployment
sphinx
sphinx_rtd_theme
nose
coverage
pypi-publisher

# Runtime
PyYAML==3.12
luigi==2.7.3
checksumdir==1.1.4
pathaliases==0.0.1
networkx==2.1
pydot==1.2.4
cookiecutter==1.6.0